reservation-service/requirements.txt
fastapi==0.95.1
uvicorn==0.21.1
httpx==0.23.3
pyjwt==2.6.0
kafka-python==2.0.2
authlib==1.3.1
starlette==0.26.1
sqlalchemy==1.4.46
python-dotenv==1.0.0
requests==2.28.2
psycopg2-binary==2.9.5  # Added this line